+++ ARK2D/src/ARK2D/Font/Font.h
@@ -7,7 +7,7 @@
 
 #ifndef FONT_H_
 #define FONT_H_
- 
+
 #include <string>
 using namespace std;
 
@@ -28,10 +28,10 @@ class FTFont;
 		class ARK2D_API ARK::Font::Font : public ARK::Core::Resource {
 			public:
 				Font();
-				virtual void drawString(const string& str, int x, int y) = 0;
-				virtual void drawString(const string& str, int x, int y, int drawz) = 0;
+				virtual void drawString(const string& str, float x, float y) = 0;
+				virtual void drawString(const string& str, float x, float y, float drawz) = 0;
 				virtual void drawStringCentered(const string& Str, int left_x, int right_x, int y);
-				virtual void drawStringCenteredAt(const string& Str, int x, int y);
+				virtual void drawStringCenteredAt(const string& Str, float x, float y);
 				virtual unsigned int getStringWidth(const string& Str) const = 0;
 				virtual unsigned int getStringHeight(const string& Str) const = 0;
 				virtual unsigned int getLineHeight() const = 0;
